requirements.txt
Django==5.2.4
python-dotenv==1.1.1
openai==1.107.1

